+++ flserver/internal/model/HireManager.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type HireManager struct {
-	ID					int64		`json:"id"`
-	AccountID 			int64		`json:"accountId"`
+	ID					int		`json:"id"`
+	AccountID 			int		`json:"accountId"`
 	RegistrationDate	time.Time	`json:"registrationDate"`
 	Location			string 		`json:"location"`
-	CompanyId			int64 		`json:"companyId"`
+	CompanyId			int 		`json:"companyId"`
 }
