+++ flserver/internal/model/InnerInfo.go
@@ -1,7 +1,7 @@
 package model
 
 type InnerInfo struct {
-	UserID int64 `json:"user_id"`
+	UserID int `json:"user_id"`
 	WhoSeeProfile string `json:"who_see_profile"`
 	ControlQuestion string `json:"control_question"`
 	ControlAnswer string `json:"-"`
