+++ flserver/internal/model/Notification.go
@@ -1,7 +1,7 @@
 package model
 
 type Notification struct {
-	UserID int64 `json:"-"`
+	UserID int `json:"-"`
 	NewMessages bool `json:"new_messages"`
 	NewProjects bool `json:"new_projects"`
 	NewsFromService bool `json:"news_service"`
