+++ flserver/internal/model/freelancer.go
@@ -3,8 +3,8 @@ package model
 import "time"
 
 type Freelancer struct {
-	ID                int64 	`json:"id"`
-	AccountId         int64 	`json:"accountId"`
+	ID                int 	`json:"id"`
+	AccountId         int 	`json:"accountId"`
 	RegistrationDate  time.Time `json:"registrationDate"`
 	Country           string	`json:"country"`
 	City              string	`json:"city"`
@@ -12,6 +12,6 @@ type Freelancer struct {
 	Phone             string	`json:"phone"`
 	TagLine           string	`json:"tagline"`
 	Overview          string	`json:"overview"`
-	ExperienceLevelId int64		`json:"experienceLevelId"`
-	SpecialityId      int64		`json:"specialityId"`
+	ExperienceLevelId int		`json:"experienceLevelId"`
+	SpecialityId      int		`json:"specialityId"`
 }
